Se arreglaron errores de sintaxis y compatibilidad con data-test.sql
