+++ src/main/java/com/example/unirideapi/model/enums/EstadoSolicitud.java
@@ -4,6 +4,5 @@ public enum EstadoSolicitud {
     PENDIENTE,
     ACEPTADO,
     RECHAZADO,
-    PENDIENTE,
     CANCELADO
 }
